pyproject.toml
[project]
name = "finance-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "dj-database-url>=2.2.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "python-dotenv>=1.2.2",
    "psycopg[binary]>=3.3.3",
    "gunicorn>=25.3.0",
    "whitenoise>=6.12.0",
]
